call renderCardsinHandLabel at beginning to get appropriate width of textLabel. Using SDL_RenderCopy of textLocation throughout the program.
Added a struct "Points" to deal with bezier curve animations

Created getCubeBezierCurve to create each point to follow a curve.

Created animateDealingCardsToPlayers that animates cards being handed out to each player

Added updates to the textTextures for when other animations were running and subsequently erasing the textTextures.

Changed the deckNumTextLocation x display location to be set to the textLocation width + a 10 unit buffer. This insures that the deckNumTextLocation is always right after textLocation

Created animateSingleCard recursive function to animate the cards in animateDealingCardsToPlayers function.